backend/requirements.txt
# ============================================
# CORE FRAMEWORK
# ============================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# ============================================
# DATABASE
# ============================================
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1

# ============================================
# LLM & AI
# ============================================
langchain==0.1.0
langgraph==0.0.20
langchain-community==0.0.10
transformers==4.33.3
torch>=2.1.0
accelerate==0.25.0
sentence-transformers==2.2.2

# Note: bitsandbytes may not work on Windows
bitsandbytes==0.41.3; sys_platform != 'win32'

# ============================================
# STT/TTS
# ============================================
faster-whisper==0.10.0
TTS==0.20.0
pydub==0.25.1
soundfile==0.12.1

# ============================================
# RESUME PARSING
# ============================================
PyPDF2==3.0.1
python-docx==1.1.0
spacy==3.7.2
pdfplumber==0.10.3

# ============================================
# AUTHENTICATION & SECURITY
# ============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ============================================
# EMAIL/SMS (OTP)
# ============================================
twilio==8.11.0
python-decouple==3.8

# ============================================
# UTILITIES
# ============================================
numpy>=1.22.0
pandas==1.5.3
scikit-learn==1.3.0
python-dateutil==2.8.2
aiofiles==23.2.1

# ============================================
# ADDITIONAL DEPENDENCIES
# ============================================
huggingface-hub>=0.19.0
tokenizers>=0.11.1
requests>=2.31.0